requirements.txt
openai>=1.0.0
python-dotenv>=1.0.1
numpy>=1.26.0
